requirements.txt
pdfminer.six
pymupdf 
easyocr
numpy
dotenv
